adding IP and email check methods in whoisxml
