+++ pyoti/domains/__init__.py
@@ -1,4 +1,3 @@
 from .checkdmarc import CheckDMARC
 from .circlpdns import CIRCLPDNS
 from .irisinvestigate import IrisInvestigate
-from .whoisxml import WhoisXML
